persistence<Repos>: Repositorios faltantes y CRUD count y save

+++ src/main/java/repositories/CrudInterface.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface CrudInterface<T> {
 
-  Integer count();
+  Long count();
 
   void save(T someEntity);
